// src/domain/service-order/service-order-status.enum.ts
export enum ServiceOrderStatus {
  RECEBIDA = 'RECEBIDA',
  EM_DIAGNOSTICO = 'EM_DIAGNOSTICO',
  AGUARDANDO_APROVACAO = 'AGUARDANDO_APROVACAO',
  EM_EXECUCAO = 'EM_EXECUCAO',
  PAUSADO = 'PAUSADO',
  FINALIZADA = 'FINALIZADA',
  ENTREGUE = 'ENTREGUE',
  ENCERRADA_SEM_EXECUCAO = 'ENCERRADA_SEM_EXECUCAO',
}